add CrudTrait to models that exist, but don't have it
When `php artisan backpack:crud article` is run, and Models\Article.php exists, but does not have CrudTrait, this change makes sure that CrudTrait is added. This makes it easy to use existing models, and just generate admin panels for them. It also allows you to generate the models with whatever tool you want (laracasts, blueprint, whatever), and then just run the command to generate an admin panel for it.