Generate CrudControllers with actual fields and columns instead of setFromDb
This PR replaces the setFromDb() statements in the CrudController stub with actual column or fields, so that it is easier for the developer then to just change the type of a field, or of a column, or remove a field or a column. Use the new fluent syntax in 4.1 since that's the one that is easiest to use for this type of changes.